+++ ChangeLog.md
@@ -12,6 +12,7 @@
     - Review async functions
     - Define new onUpdateRegisterFn() to be called on each field update at Checker level
     - Remove IInstanciationArgs and IEnumerable interfaces, consolidating them in Panel class
+    - Checker.getForm() and Checker.setForm() are moved to Panel class
 
 ### 1.5.0